move fdr to fdr_
