+++ alphadia/_fdrx/models/two_step_classifier.py
@@ -6,8 +6,8 @@ import logging
 import numpy as np
 import pandas as pd
 
-from alphadia.fdr import get_q_values, keep_best
 from alphadia.fdr_.classifiers import Classifier
+from alphadia.fdr_.fdr import get_q_values, keep_best
 
 logger = logging.getLogger()
