+++ alphadia/workflow/managers/fdr_manager.py
@@ -10,8 +10,8 @@ import torch
 import xxhash
 
 import alphadia
-from alphadia import fdr
 from alphadia._fdrx.models.two_step_classifier import TwoStepClassifier
+from alphadia.fdr_ import fdr
 from alphadia.workflow.managers.base import BaseManager
 
 logger = logging.getLogger()
