+++ tests/performance_tests/fdr_test.py
@@ -13,7 +13,7 @@ import torch
 from alphabase.tools.data_downloader import DataShareDownloader
 
 import alphadia
-from alphadia import fdr
+from alphadia.fdr_ import fdr
 from alphadia.fdr_.classifiers import BinaryClassifierLegacyNewBatching
 from alphadia.workflow.peptidecentric.peptidecentric import feature_columns
